+++ packages/wrapper/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/package/monaco-editor-wrapper) are documented in this file.
 
-## [6.0.0-next.5] - 2024-10-23
+## [6.0.0-next.6] - 2024-10-30
 
 - Moved `createUrl` to `monaco-languageclient/tools`, moved editor app specific utils to editorAppbase and create function `didModelContentChange`.
 - Added
@@ -11,7 +11,7 @@ All notable changes to npm module [monaco-editor-wrapper](https://www.npmjs.com/
 - Support all arguments for monaco-vscode-api `initialize` [#756](https://github.com/TypeFox/monaco-languageclient/pull/756)
   - This also allows to configure editor-, view- or workspace-service. This is a preparation for further enhancements.
 - Update to monaco-vscode-api 9.0.x [#749](https://github.com/TypeFox/monaco-languageclient/pull/749)
-  - Updated to `monaco-languageclient@9.0.0-next.5`. Updated all `@codingame/monaco-vscode` packages to `10.1.1`.
+  - Updated to `monaco-languageclient@9.0.0-next.6`. Updated all `@codingame/monaco-vscode` packages to `10.1.3`.
   - Enhancements to logging, worker factory and start order. The worker factory only accepts direct worker loading instructions from now on
 - monaco-languageclient config improvement, wrapper+languageclientwrapper improvements [#741](https://github.com/TypeFox/monaco-languageclient/pull/741)
   - Allows to configure more than one language client